Correction de quasi tous les warnings
Il reste seulement le warning C26812, qui apparait à des endroits différents à chaque fois...

+++ Source/Entity/Entity.cpp
@@ -1,7 +1,7 @@
 #include "Entity.h"
 
 
-Entity::Entity(std::string texturePath, int defaultLife, sf::Vector2f initPosition) //Constructeur par défaut, sans paramètre
+Entity::Entity(std::string texturePath, float defaultLife, sf::Vector2f initPosition) //Constructeur par défaut, sans paramètre
 {
 	//A la création d'un nouveau joueur, on lui attribue des caractéristiques:
