+++ Source/Entity/NPC/Ennemy.cpp
@@ -1,6 +1,6 @@
 #include "Ennemy.h"
 
-Ennemy::Ennemy(std::string texturePath, int defaultLife, sf::Vector2f initPosition) : NPC(texturePath, defaultLife, initPosition) {
+Ennemy::Ennemy(std::string texturePath, float defaultLife, sf::Vector2f initPosition) : NPC(texturePath, defaultLife, initPosition) {
 	this->setWeapon(Arme("Time-Quest/Source/assets/mas38final.png", 5));
 }
