+++ Source/Entity/NPC/Ennemy.h
@@ -4,7 +4,7 @@
 class Ennemy : public NPC
 {
 public:
-	Ennemy(std::string texturePath, int defaultLife, sf::Vector2f initPosition);
+	Ennemy(std::string texturePath, float defaultLife, sf::Vector2f initPosition);
 	~Ennemy();
 
 	void update(sf::Vector2f playerPos);
