+++ Source/Entity/NPC/NPC.cpp
@@ -1,6 +1,6 @@
 #include "NPC.h"
 
-NPC::NPC(std::string texturePath, int defaultLife, sf::Vector2f initPosition) : Entity(texturePath, defaultLife, initPosition) {
+NPC::NPC(std::string texturePath, float defaultLife, sf::Vector2f initPosition) : Entity(texturePath, defaultLife, initPosition) {
 	_lifeBar.setFillColor(sf::Color::Red);
 	_lifeBar.setOutlineThickness(1);
 }
