+++ Source/Entity/Player.h
@@ -4,7 +4,7 @@
 class Player : public Entity
 {
 public:
-	Player(std::string texturePath, int defaultLife, sf::Vector2f initPosition);
+	Player(std::string texturePath, float defaultLife, sf::Vector2f initPosition);
 	~Player();
 
 	sf::Vector2f getPosition() const;
